Adding a "wars" dir to store the app compressed.
Adding to the pom servlet dependency (3.1)
and the plugin for cleaning the target dir when using the package phase (mvn package).
This is because when using "mvn package" the /target dir is not being removed by default.
As we're developing a webapp here we don't need the target dir but the actual /war.